password change

+++ app/Http/Controllers/AdminDashboard.php
@@ -15,7 +15,6 @@ class AdminDashboard extends Controller
     public function AdminDashboard()
     {
         return view('admin.index');
-        
     }
 
     public function AdminLogout(Request $request): RedirectResponse
@@ -34,13 +33,15 @@ class AdminDashboard extends Controller
         return view('admin.admin_login');
     }
 
-    public function AdminProfile(){
+    public function AdminProfile()
+    {
         $id = Auth::user()->id;
         $profileData = User::findOrFail($id);
         return view('admin.admin_profile', compact('profileData'));
     }
 
-    public function AdminProfileUpdate(Request $request){
+    public function AdminProfileUpdate(Request $request)
+    {
 
         $request->validate([
             'email' => 'email',
@@ -51,7 +52,7 @@ class AdminDashboard extends Controller
         $profileData = User::findOrFail($id);
 
         $username = $request->username;
-        $username_lower = Str::lower($username); 
+        $username_lower = Str::lower($username);
         $username_remove_space = Str::replace(' ', '', $username_lower);
 
         User::find($id)->update([
@@ -63,10 +64,12 @@ class AdminDashboard extends Controller
         ]);
 
         if ($request->hasFile('photo')) {
-            unlink(base_path('public/upload/admin_images/'.$profileData->photo));
+            if ($profileData->photo) {
+                unlink(base_path('public/upload/admin_images/' . $profileData->photo));
+            }
             $file_name = auth()->id() . time() . '.' . $request->file('photo')->getClientOriginalExtension();
             $img = Image::make($request->file('photo'));
-            $img->save(base_path('public/upload/admin_images/'.$file_name));
+            $img->save(base_path('public/upload/admin_images/' . $file_name));
             User::find($id)->update([
                 'photo' => $file_name,
             ]);
@@ -80,55 +83,49 @@ class AdminDashboard extends Controller
         return back()->with($notification);
     }
 
-    public function AdminChangePassword(){
+    public function AdminChangePassword()
+    {
         $id = Auth::user()->id;
         $profileData = User::findOrFail($id);
         return view('admin.admin_change_password', compact('profileData'));
     }
 
-    public function AdminUpdatePassword(Request $request){
+    public function AdminUpdatePassword(Request $request)
+    {
         $id = Auth::user()->id;
         $request->validate([
             'old_password' => 'required',
-            'new_password' => 'required|confirmed',
+            'new_password' => 'required|confirmed|min:6',
         ]);
 
         if (!Hash::check($request->old_password, auth()->user()->password)) {
-            
+
             $notification = array(
                 'message' => 'Old Password Does Not Match',
                 'alert-type' => 'error',
             );
 
             return back()->with($notification);
-        } 
-        // else {
-        //     if ($request->old_password === $request->new_password) {
-        //         $notification = array(
-        //             'message' => 'Old password and new password same!',
-        //             'alert-type' => 'error',
-        //         );
-    
-        //         return back()->with($notification);
-        //     } else {
-                
-        //     }
-            
-            
-        // }
+        } else {
+            if ($request->old_password === $request->new_password) {
+                $notification = array(
+                    'message' => 'Old password and new password same!',
+                    'alert-type' => 'error',
+                );
 
-        User::find($id)->update([
-            'password' => Hash::make($request->new_password),
-        ]);
+                return back()->with($notification);
+            } else {
+                User::find($id)->update([
+                    'password' => Hash::make($request->new_password),
+                ]);
 
-        $notification = array(
-            'message' => 'Password Changed Successfully',
-            'alert-type' => 'success',
-        );
+                $notification = array(
+                    'message' => 'Password Changed Successfully',
+                    'alert-type' => 'success',
+                );
 
-        return back()->with($notification);
-     
-
-        
+                return back()->with($notification);
+            }
+        }
     }
 }
